blogs bugs fixed

+++ src/main/java/com/oeps/onlineeventplanningsystem/controller/BlogController.java
@@ -1,5 +1,6 @@
 package com.oeps.onlineeventplanningsystem.controller;
 
+import com.oeps.onlineeventplanningsystem.error.BlogsExceptions;
 import com.oeps.onlineeventplanningsystem.model.Blog;
 import com.oeps.onlineeventplanningsystem.model.User;
 import com.oeps.onlineeventplanningsystem.repositories.BlogRepo;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.util.*;
 import java.util.List;
 import java.util.Optional;
@@ -25,16 +27,25 @@ public class BlogController {
     BlogRepo blogRepo;
 
     @GetMapping("/addNewBlog")
-    public String addBlog(String blogTitle, String author, String blogDescription , String blogImage,String blogContent){
+    public String addBlog(String blogTitle, String author, String blogDescription , String blogImage, String blogContent , HttpSession session) throws BlogsExceptions {
         Blog blog = new Blog();
+        String username = ((User) session.getAttribute("userSession")).getUsername();
 
-        blog.setBlogTitle(blogTitle);
-        blog.setAuthor(author);
-        blog.setBlogDescription(blogDescription);
-        blog.setBlogImage(blogImage);
-        blog.setBlogContent(blogContent);
+        if (blogImage.length() > 3500){
 
-        blogRepo.save(blog);
+            throw new BlogsExceptions("Image size should be less than 3500 bytes");
+        }else {
+            blog.setBlogTitle(blogTitle);
+            blog.setAuthor(author);
+            blog.setBlogDescription(blogDescription);
+            blog.setBlogImage(blogImage);
+
+            blog.setBlogContent(blogContent);
+            blog.setUserName(username);
+
+            blogRepo.save(blog);
+        }
+
 
         return "redirect:/blogs";
     }
@@ -86,6 +97,17 @@ public class BlogController {
         },HttpStatus.OK);
     }
 
+    @GetMapping("/deleteBlog/{id}")
+    public ModelAndView getBlogId(@PathVariable("id") int blogID){
+        Blog blog = blogRepo.findById(blogID).get();
+        return new ModelAndView("/deleteBlog", new HashMap(){
+            {
+                put("blogID" , blog);
+            }
+
+        },HttpStatus.OK);
+    }
+
 
     @PostMapping("/updateBlog/newUpdatedBlog/{blogID}")
     public String getNewUpdatedBlogs(@PathVariable("blogID") int blogID ,String blogTitle, String author, String blogDescription , String blogImage,String blogContent){
@@ -106,10 +128,13 @@ public class BlogController {
 
 
 
-    @GetMapping("/deleteExistingBlog")
-    public String deleteBlog(int blogID){
+    @PostMapping("/deleteBlog/deleteExistingBlog")
+    public String deleteBlog(int blogID) throws BlogsExceptions {
 
         Optional<Blog> blog= blogRepo.findById(blogID);
+        if (blog.isEmpty()){
+            throw new BlogsExceptions("Blog not found with id: " + blogID);
+        }
 
         if(blog.isPresent()) {
             blogRepo.delete(blog.get());
